Compile

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -12,15 +12,15 @@ return (0);
 void print_fibonacci(int n)
 {
 int i;
-long long int fib1 = 1;
-long long int fib2 = 2;
-long long int next;
+unsigned long fib1 = 1;
+unsigned long fib2 = 2;
+unsigned long next;
 
-printf("%lld, %lld", fib1, fib2);
+printf("%lu, %lu", fib1, fib2);
 for (i = 3; i <= n; i++)
 {
 next = fib1 + fib2;
-printf(", %lld", next);
+printf(", %lu", next);
 fib1 = fib2;
 fib2 = next;
 }
